Fix possible invalid history after form submit
In Chrome 40, when submitting a form and quickly clicking back twice the URL will point to the page before the form while the form is being shown.
This is due to the delay to to bypass the initial load state (e69bead), setting this delay to 0 fixes the issue and keeps the same functionality.