Don't check turbolinks variable in popstate event
If a third party library modifies the history, the page won't be initialized correctly when you return to it the first time.

As a (very) basic example, using a console on the test server: history.pushState({}, 'test', 'http://localhost:9292/other.html'); > changes the URL, history.back(); > goes back to the previous page, history.forward(); > go back to the last page.
After the forward the popstate event is ignored by turbolinks since there is no turbolinks variable in the event.state object. This results in the page not being initialized correctly; any javascript required for the page won't be executed.